chore: Clean up artwork rail and extact sub components
